refactor(admin): AI triage via Supabase Edge Function (remove Next route)
- Add functions/admin-challenge-ai-review + shared CORS
- Client calls /functions/v1 with session JWT + anon apikey
- Secrets OPENAI_* live on Supabase only; exclude functions from tsc
