+++ .env.local.example
@@ -11,10 +11,12 @@ NEXT_PUBLIC_SITE_URL=https://truegrynd.vercel.app
 # Used for admin operations (seeding, RLS bypass)
 # SUPABASE_SERVICE_ROLE_KEY=eyJhbGci...
 
-# Admin AI triage (Route Handler only — never NEXT_PUBLIC_*)
-# OPENAI_API_KEY=sk-...
-# OPENAI_MODEL=gpt-4o-mini
-
 # E2E auth test user (Supabase auth.users)
 E2E_TEST_EMAIL=test@example.com
 E2E_TEST_PASSWORD=testpassword123
+
+# --- Admin AI triage (Supabase Edge Function, not Vercel) ---
+# Deploy: supabase functions deploy admin-challenge-ai-review
+# Secrets (Supabase Dashboard → Project → Edge Functions → Secrets):
+#   OPENAI_API_KEY=sk-...
+#   OPENAI_MODEL=gpt-4o-mini   # optional; default in function is gpt-4o-mini
