+++ src/features/admin/services/adminChallenges.ts
@@ -93,22 +93,32 @@ export async function adminBatchApproveChallenges(
   return typeof data === 'number' ? data : Number(data);
 }
 
-export async function adminRunChallengeAiReview(input: {
-  challengeId: string;
-  accessToken: string;
-}): Promise<void> {
-  const res = await fetch(
-    `/api/admin/challenges/${encodeURIComponent(input.challengeId)}/ai-review`,
-    {
-      method: 'POST',
-      headers: {
-        Authorization: `Bearer ${input.accessToken}`,
-      },
-    },
-  );
+export async function adminRunChallengeAiReview(input: { challengeId: string }): Promise<void> {
+  const {
+    data: { session },
+  } = await supabase.auth.getSession();
+  if (!session?.access_token) {
+    throw new Error('no_session');
+  }
 
+  const baseUrl = process.env.NEXT_PUBLIC_SUPABASE_URL?.replace(/\/$/, '') ?? '';
+  const anonKey = process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY ?? '';
+  if (!baseUrl || !anonKey) {
+    throw new Error('server_misconfigured');
+  }
+
+  const res = await fetch(`${baseUrl}/functions/v1/admin-challenge-ai-review`, {
+    method: 'POST',
+    headers: {
+      'Content-Type': 'application/json',
+      Authorization: `Bearer ${session.access_token}`,
+      apikey: anonKey,
+    },
+    body: JSON.stringify({ challenge_id: input.challengeId }),
+  });
+
+  const parsed = (await res.json().catch(() => ({}))) as { code?: string };
   if (!res.ok) {
-    const parsed = (await res.json().catch(() => null)) as { code?: string } | null;
-    throw new Error(parsed?.code ?? `http_${res.status}`);
+    throw new Error(parsed.code ?? `http_${res.status}`);
   }
 }
