+++ supabase/functions/_shared/cors.ts
@@ -1,0 +1,4 @@
+export const corsHeaders: Record<string, string> = {
+  'Access-Control-Allow-Origin': '*',
+  'Access-Control-Allow-Headers': 'authorization, x-client-info, apikey, content-type',
+};
